removed user-data from fastn_ds
