+++ fastn-core/src/auth/ud.rs
@@ -1,1 +1,8 @@
-pub use fastn_ds::ud::UserData;
+#[derive(serde::Deserialize, Clone)]
+pub struct UserData {
+    pub id: i64,
+    pub username: String,
+    pub name: String,
+    pub email: String,
+    pub verified_email: bool,
+}
